requirements.txt
asyncio==3.4.3
sqlalchemy==1.4.45
python-dotenv==0.21.0
asyncpg==0.27.0
psycopg2==2.9.5
aiohttp==3.8.3
